Use echo and pipe to get upload data into curl
When using the script, I was seeing "Argument list too long" errors with the `curl` command, presumably because of the length of `$upload`.  Using `echo` to get the content of `$upload` and piping it into `curl` with `-d @-` to read data from the pipe works in the case where I had been getting the "Argument list too long" error.