Implementing parent picking
This PR has the intention to allow the commit parent to be picked on client side.

Usually, the parent is determined by the system by looking on Github/Bitbucket/Gitlab for the parents of the commit.
When those are not on Codecov database, the system keeps searching among the ancestors for the commit that is the closest ancestor of the given commit. This is helpful because not everyone uploads all commits to codecov, specially when we are doing some long work with multiple commits and CI is applied only to the latest one.

This behavior automatic and will keep existing. But there are some situations where we know for a fact what the parent commit should be exactly, and would rather force that to be used than let the system determine it with the information it has right now.

One good example is when the upload order of the commits is for some reason reversed. If A->B->C are commits, and C first report upload happens before B first upload, the system won't know that commit B is a good commit parent for C when C gets uploaded, because B is not in the database yet. Then A will be set as C parent. But if we know B is eventually going to be uploaded soon, being able to set that B will be the parent is very helpful.